requirements.txt
tensorflow>=2.16.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
numpy>=1.26.0
Pillow>=10.0.0
